zcash_pool_migration: share the migration-state test fixtures
`state` and `satisfiability` each carried the same six fixtures, and
`satisfiability` said so in a comment: "The state constructors mirror
`state.rs`'s test helpers, so a drive-loop test and a kernel test describe
the same migration the same way."

Five of the six were byte-identical, doc comments included. The sixth,
`state_with_crossings`, differed only in that `satisfiability`'s copy had
inlined the fee-buffer parameter as a literal, so the shared version keeps
the parameter and its call sites pass what the copy hardcoded.

They move to `testing::fixtures`, next to the fixtures gathered there in
the previous commit. They construct `MigrationTransaction` and
`MigrationState` by field, so they could not live in the memory crate at
any visibility; a module inside this crate is the only home that works.

`testing::fixtures` is now gated on `cfg(test)` alone rather than also on
`orchard`, since these six need neither the feature nor the memory crate.
The parts that do need it are gated individually.
