+++ zcash_pool_migration/src/testing/mod.rs
@@ -21,7 +21,7 @@
 //! means, and `scenarios` holds the fixed data both of the above speak about.
 
 mod conformance;
-#[cfg(all(test, feature = "orchard"))]
+#[cfg(test)]
 pub(crate) mod fixtures;
 mod scenarios;
 mod strategies;
